+++ gradle-test-extensions-subprojects/gradle-testkit-test-extensions/main/kotlin/com/javiersc/gradle/testkit/test/extensions/resourceFile.kt
@@ -1,6 +1,0 @@
-package com.javiersc.gradle.testkit.test.extensions
-
-import java.io.File
-
-public fun resourceFile(resource: String): File =
-    File(Thread.currentThread().contextClassLoader?.getResource(resource)?.toURI()!!)