+++ integration-tests/test/kotlin/com/javiersc/integration/tests/GradleProjectTestExtensionsTest.kt
@@ -1,18 +1,18 @@
 package com.javiersc.integration.tests
 
-import com.javiersc.gradle.testkit.test.extensions.GradleTest
+import com.javiersc.gradle.project.test.extensions.GradleProjectTest
 import io.kotest.matchers.shouldBe
 import kotlin.test.Test
 
-internal class GradleTestExtensionsTest : GradleTest() {
+internal class GradleProjectTestExtensionsTest : GradleProjectTest() {
 
     @Test
     fun `GIVEN a null sandbox path WHEN project builds THEN project's name is an empty string`() {
-        gradleTest(sandboxPath = null) { name.shouldBe("") }
+        gradleProjectTest(sandboxPath = null) { name.shouldBe("") }
     }
 
     @Test
     fun `GIVEN a sandbox path WHEN project builds THEN project's name is 'fake-file'`() {
-        gradleTest(sandboxPath = "fake-file") { name.shouldBe("fake-file") }
+        gradleProjectTest(sandboxPath = "fake-file") { name.shouldBe("fake-file") }
     }
 }