+++ integration-tests/test/kotlin/com/javiersc/integration/tests/GradleTestKitExtensionsTest.kt
@@ -1,11 +1,10 @@
 package com.javiersc.integration.tests
 
-import com.javiersc.gradle.testkit.test.extensions.GradleTest
-import com.javiersc.gradle.testkit.test.extensions.resourceFile
+import com.javiersc.gradle.testkit.test.extensions.GradleTestKitTest
 import io.kotest.matchers.shouldBe
 import kotlin.test.Test
 
-internal class GradleTestKitExtensionsTest : GradleTest() {
+internal class GradleTestKitExtensionsTest : GradleTestKitTest() {
 
     @Test
     fun test() {